core/tawala/pyproject.toml
[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"

[project.urls]
Repository = "https://github.com/treeolivetech/pkg-tawala"
Issues = "https://github.com/treeolivetech/pkg-tawala/issues"

[project.scripts]
tawala = "tawala.management.cli:main"
twl = "tawala.management.cli:main"

[project]
name = "tawala"
version = "0.1.0"
description = "Build and deploy Django apps with confidence."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Kevin Wasike Wakhisi", email = "kevin@treeolive.tech" }]
requires-python = ">=3.14"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Framework :: Django :: 6",
    "Typing :: Typed",
]
dependencies = [
    "django-browser-reload==1.21.0",
    "django-compressor==4.6.0",
    "django-http-compression[brotli]==1.4.0",
    "django-minify-html==1.14.0",
    "pyperclip==1.11.0",
    "tawala-ui==0.1.0",
]

[project.optional-dependencies]
wip = ["tawala-wip==0.1.0"]
school = ["tawala-school==0.1.0"]

[tool.uv.sources]
tawala-ui = { workspace = true }
tawala-wip = { workspace = true }
tawala-school = { workspace = true }
